refactor: extract shared utilities and fix team-edge position bug
- Extract markdownToHtml to src/lib/utils.ts (removes 3 duplicates)
- Extract ROLE_COLORS to src/lib/utils.ts (removes 2 duplicates)
- Fix team-edge.tsx bug where labelX/labelY were undefined in onMutate
